--- xcq-types/tests/main.rs
@@ -1,4 +1,0 @@
-#[macro_use]
-mod macros;
-mod derive;
-mod prelude;
